split bin.c added signal handling.
