+++ sources/include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: minsunki <minsunki@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/02 11:56:27 by minsunki          #+#    #+#             */
-/*   Updated: 2022/03/08 17:13:42 by minsunki         ###   ########seoul.kr  */
+/*   Updated: 2022/03/08 21:15:18 by minsunki         ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -82,6 +82,7 @@ typedef struct s_meta
 	int			pid;
 	int			child;
 	int			hd_cur;
+	int			waiting;
 }				t_meta;
 
 typedef unsigned char	t_byte;
@@ -127,8 +128,10 @@ void	heredoc_read(t_meta *m, char *tstr, int fd);
 void	exec_start(t_meta *m);
 
 // bin.c
-char	*bin_find(t_meta *m, char *bin);
 int		bin_run(t_meta *m, char *bin);
+
+// bin_util.c
+char	*bin_find(t_meta *m, char *bin);
 
 // argv.c
 char	**argv_build(t_token *tok);
