+++ sources/signal.c
@@ -6,23 +6,27 @@
 /*   By: minsunki <minsunki@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/01 13:23:14 by tjung             #+#    #+#             */
-/*   Updated: 2022/03/02 23:41:16 by minsunki         ###   ########seoul.kr  */
+/*   Updated: 2022/03/08 21:17:06 by minsunki         ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-// 자식 프로세스는 자식 handler 등록 계획.
+//ctrl-\ = 3 ctrl-c = 2
 
 void	sig_handler(int signum)
 {
-	if (signum != SIGINT)
-		return ;
+	t_meta *m;
+
+	m = meta_get();
 	write(STDOUT_FILENO, "\n", 1);
-	if (rl_on_new_line() == -1)
-		exit(1);
-	rl_replace_line("", 0);
-	rl_redisplay();
+	if (!m->waiting && signum == SIGINT)
+	{
+		if (rl_on_new_line() == -1)
+			mexit(1);
+		rl_replace_line("", 0);
+		rl_redisplay();
+	}
 }
 
 void	set_signal(void)
